requirements.txt
flask
requests
beautifulsoup4
nltk
faiss-cpu
scikit-learn
openai==0.28
